Disabled VOWL in Widoco (load time too slow)

+++ publisher/product/widoco/build.sh
@@ -210,8 +210,6 @@ function generateWidocoDocumentationForFile() {
   #
   cp "${ontologyPolicyFile}" .
 
-  java -version 2>&1 | pipelog
-
   java \
     -classpath /usr/share/java/log4j/log4j-core.jar:/usr/share/java/log4j/log4j-1.2-api.jar:/usr/share/java/log4j/log4j-api.jar \
     -Dxxx=widoco \
@@ -229,8 +227,7 @@ function generateWidocoDocumentationForFile() {
     -doNotDisplaySerializations \
     -displayDirectImportsOnly \
     -lang en  \
-    -getOntologyMetadata \
-    -webVowl 2>&1 | \
+    -getOntologyMetadata 2>&1 | \
     grep -v "JenaCatalogIRIMapper.* -> " | \
     grep -v 'WIzard' | \
     grep -v 'https://w3id.org/widoco/' | \
@@ -273,12 +270,15 @@ function generateWidocoDocumentationForFile() {
   #fi
 
   #
-  # Remove the default ontologies that come with WebVowl
+  # If webvowl output was generated,
+  # remove the default ontologies that come with WebVowl
   #
-  (
-    cd "${outputDir}/webvowl/data" || return $?
-    rm -vf foaf.json goodrelations.json muto.json new_ontology.json ontovibe.json personasonto.json sioc.json template.json
-  )
+  if [[ -d "${outputDir}/${rdfFileNoExtension}/webvowl" ]] ; then
+    (
+      cd "${outputDir}/${rdfFileNoExtension}/webvowl/data" || return $?
+      rm -vf foaf.json goodrelations.json muto.json new_ontology.json ontovibe.json personasonto.json sioc.json template.json
+    )
+  fi
 
   return 0
 }
